requirements.txt
PyDispatcher
eyeD3
pygame
gTTS
SpeechRecognition
Metaphone
requests
psutil
